+++ src/main/java/com/damian/EmailConstants.java
@@ -1,8 +1,10 @@
 package com.damian;
 
 public class EmailConstants {
-    public static final int LOGIN_STATE_NOT_READY = 0;
-    public static final int LOGIN_STATE_FAILED_BY_NETWORK = 1;
-    public static final int LOGIN_STATE_FAILED_BY_CREDENTIALS = 2;
-    public static final int LOGIN_STATE_SUCCEDED = 3;
+	
+	public static final int LOGIN_STATE_NOT_READY = 0;
+	public static final int LOGIN_STATE_FAILED_BY_NETWORK = 1;
+	public static final int LOGIN_STATE_FAILED_BY_CREDENTIALS = 2;
+	public static final int LOGIN_STATE_SUCCEDED = 3;
+
 }
