+++ src/main/java/com/damian/Main.java
@@ -5,8 +5,6 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 
-
-
 public class Main extends Application {
 
     public static void main(String[] args) {
