+++ src/main/java/com/damian/ViewFactory.java
@@ -5,7 +5,6 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 
-
 import java.io.IOException;
 
 /**
