+++ src/main/java/com/damian/model/Singleton.java
@@ -1,7 +1,5 @@
 package com.damian.model;
 
-import com.damian.EmailMessageBean;
-
 public class Singleton {
 
     private Singleton() {}
